refactor(runner): extract agent-task provider preflight from lab offload god-module to harden API boundary (#4303)
